03_cloud_functions/requirements.txt
google-cloud-datastore==1.7.1